Move the coordinator logic to the BroadlinkEntity class
